refactor: move change auth controller to infra folder
